:art: Moved history item to utils for modularity
